requirements.txt
streamlit==1.37.1
torch==2.2.2
torchvision==0.17.2
scikit-learn==1.5.1
pillow>=10.3.0
transformers==4.42.4
numpy>=1.26.4

